* [Platform/Plugins] Fixed an bug in the DevblocksPlatform::sortObjects() method that forced all numbers to be compared as integers, thus rendering it incapable of properly sorting decimals (e.g. 0.5 and 0.3 were both rounded to zero). Numbers are now properly compared as floating point values.

+++ libs/devblocks/api/Extension.php
@@ -1199,8 +1199,8 @@ class _DevblocksSortHelper {
 		}
 		
 		if(is_numeric($a_test) && is_numeric($b_test)) {
-			settype($a_test, 'integer');
-			settype($b_test, 'integer');
+			settype($a_test, 'float');
+			settype($b_test, 'float');
 			
 			if($a_test==$b_test)
 				return 0;
